Version 2.3.1

+++ contrib/hwloc/contrib/ci.inria.fr/job-3-mingw.sh
@@ -1,8 +1,14 @@
 #!/bin/sh
 #
-# Copyright © 2012-2015 Inria.  All rights reserved.
+# Copyright © 2012-2018 Inria.  All rights reserved.
 # See COPYING in top-level directory.
 #
+
+echo "############################"
+echo "Running on:"
+uname -a
+echo "Options: $0"
+echo "############################"
 
 set -e
 set -x
@@ -44,22 +50,27 @@ while test $# -gt 0; do
     echo "  --no-install  Don't install"
     echo "  --help        Show this help"
     exit 0
+  else
+    break
   fi fi fi fi fi fi fi fi fi
   shift
 done
 
+oldPWD=$PWD
 oldPATH=$PATH
 
-if test x$dotar = x1; then
-  # remove everything but the last 10 builds
-  rm -rf $(ls | grep -v ^hwloc- | grep -v ^job-) || true
-  chmod u+w -R $(ls -td hwloc-* | tail -n +11) || true
-  rm -rf $(ls -td hwloc-* | tail -n +11) || true
+# remove previous artifacts so that they don't exported again by this build
+rm -f hwloc-win*-build-*.zip || true
 
-  # find the tarball and extract it
-  tarball=$(ls -tr hwloc-*.tar.gz | tail -1)
+if test x$dotar = x1; then
+  # extract the tarball
+  tarball="$1"
+  if ! test -f "$tarball"; then
+    echo "Invalid tarball parameter."
+    exit 0
+  fi
   basename=$(basename $tarball .tar.gz)
-  version=$(echo $basename | cut -d- -f2)
+  version=$(echo $basename | cut -d- -f2-)
   test -d $basename && chmod -R u+rwX $basename && rm -rf $basename
   tar xfz $tarball
   if test x$dokeeptar = x0; then
@@ -72,14 +83,22 @@ fi
 
 if test x$dobuild32 = x1; then
 
+  echo "*** Switching to MinGW32 ***"
+  set +xe # cannot keep set -e while sourcing /etc/profile
+  MSYSTEM=MINGW32 . /etc/profile || true
+  set -xe
+  cd $oldPWD
+
+  # for MS 'lib' program in dolib.c
+  export PATH=$PATH:$MSLIB_PATH
+
   mkdir ${basename}/build32 || true
   cd ${basename}/build32
 
-  export PATH=/c/Builds:/c/Builds/mingw32/bin/:/c/Builds/mingw64/bin/:/c/Builds/mingw32/i686-w64-mingw32/lib:"/c/Program Files (x86)/Microsoft Visual Studio 11.0/VC/bin":"/c/Program Files (x86)/Microsoft Visual Studio 11.0/Common7/IDE":$oldPATH
   if test x$doconf = x1; then
     winball=hwloc-win32-build-${version}
     prefix=${PWD}/../${winball}
-    ../configure --prefix=$prefix --enable-static --host=i686-w64-mingw32 CC_FOR_BUILD=x86_64-w64-mingw32-gcc $confopts
+    ../configure --prefix=$prefix --enable-static CC="gcc -static-libgcc" $confopts
   fi
 
   make
@@ -101,6 +120,7 @@ if test x$dobuild32 = x1; then
   fi
 
   build32/utils/lstopo/lstopo-no-graphics -v
+  build32/utils/hwloc/hwloc-info --support
 
   if test x$dotar = x1; then
     cd ..
@@ -111,14 +131,22 @@ fi
 
 if test x$dobuild64 = x1; then
 
+  echo "*** Switching to MinGW64 ***"
+  set +xe # cannot keep set -e while sourcing /etc/profile
+  MSYSTEM=MINGW64 . /etc/profile || true
+  set -xe
+  cd $oldPWD
+
+  # for MS 'lib' program in dolib.c
+  export PATH=$PATH:$MSLIB_PATH
+
   mkdir ${basename}/build64 || true
   cd ${basename}/build64
 
-  export PATH=/c/Builds:/c/Builds/mingw64/bin/:/c/Builds/mingw32/i686-w64-mingw32/lib/:"/c/Program Files (x86)/Microsoft Visual Studio 11.0/VC/bin":"/c/Program Files (x86)/Microsoft Visual Studio 11.0/Common7/IDE":$oldPATH
   if test x$doconf = x1; then
     winball=hwloc-win64-build-${version}
     prefix=${PWD}/../${winball}
-    ../configure --prefix=$prefix --enable-static --host=x86_64-w64-mingw32 $confopts
+    ../configure --prefix=$prefix --enable-static CC="gcc -static-libgcc" $confopts
   fi
 
   make
@@ -140,6 +168,7 @@ if test x$dobuild64 = x1; then
   fi
 
   build64/utils/lstopo/lstopo-no-graphics -v
+  build64/utils/hwloc/hwloc-info --support
 
   if test x$dotar = x1; then
     cd ..
@@ -148,3 +177,5 @@ if test x$dobuild64 = x1; then
 fi
 
 PATH=$oldPATH
+
+exit 0
